Moved run/dryRun specific parameters and code from AbstractRewriteMojo into new AbstractRewriteBaseRunMojo abstract class.
The extension path for the Abstract Run and Dry Run Mojos is now:
AbstractRewriteRunMojo/AbstractRewriteDryRunMojo -> AbstractRewriteBaseRunMojo -> AbstractRewriteMojo
